fix(main): refactor test to deploy
